% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SESraster.R
\name{algorithm_performance}
\alias{algorithm_performance}
\title{Performance of randomization algorithms}
\usage{
algorithm_performance(
  x,
  algorithm,
  alg_args = NULL,
  aleats = 10,
  plot = F,
  ...
)
}
\arguments{
\item{x}{SpatRaster. A SpatRaster containing presence-absence data (0 or 1)
for a set of species.}

\item{algorithm}{character. A character indicating the randomization method.}

\item{alg_args}{List of arguments passed to the randomization method chosen
in 'algorithm'.
See \code{\link{bootspat_naive}}, \code{\link{bootspat_str}}, \code{\link{bootspat_ff}}}

\item{aleats}{positive integer. A positive integer indicating how many times
the calculation should be repeated.}

\item{plot}{logical. Should results be plotted?}

\item{...}{additional arguments passed to 'terra::app()' function.}
}
\value{
a list with three components.
\itemize{
\item randomization_results: a matrix with frequency of species occurrence on
each randomization
\item spp_metrics: a matrix with metrics comparing actual and randomized frequency
of species occurrence. Metrics are average, sd, min, and max frequency across
randomizations, sp_reldiff (average difference relative to species frequency),
global_reldiff (average difference relative to the number of available cells),
upper and lower confidence intervals for sp_reldiff and global_reldiff.
}

#' @seealso \code{\link{bootspat_str}}, \code{\link{bootspat_naive}},
\code{\link{bootspat_ff}}, \code{\link{SESraster}}
}
\description{
Compares the richness and occurrence incidence across species
between actual and randomized species distributions
}
\examples{
library(SESraster)
library(terra)
r <- load_ext_data()
appmean <- function(x, ...){
terra::app(x, "mean", ...)
}
algorithm_performance(r, FUN=appmean, algorithm = "SIM1", alg_args=list(random="species"))
algorithm_performance(r, FUN=appmean, algorithm = "SIM1", alg_args=list(random="site"))
#' algorithm_performance(r, FUN=appmean, algorithm = "SIM1", alg_args=list(random="both"))
}
\author{
Neander M. Heming
}
