% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_check.R
\name{.fit.memory}
\alias{.fit.memory}
\title{Function to evaluate if the rasters generated in the function fits on memory}
\usage{
.fit.memory(x)
}
\arguments{
\item{x}{number of rasters generated in the function}
}
\value{
logical
}
\description{
Function to evaluate if the rasters generated in the function fits on memory
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
\keyword{internal}
