% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspat_naive.R
\name{.lyr.sample}
\alias{.lyr.sample}
\title{Internal function to resample a vector according to the observed frequency}
\usage{
.lyr.sample(x, fr)
}
\arguments{
\item{x}{numeric. A vector containing values to resampling.}

\item{fr}{data.frame A data.frame with 3 columns (layer, value, count).}
}
\value{
vector
}
\description{
Internal function to resample a vector according to the observed frequency
}
\author{
Neander Marcel Heming
}
\keyword{internal}
