% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspat_naive.R
\name{.sample.not.NA}
\alias{.sample.not.NA}
\title{Intern function to sample vectors with non-NA values}
\usage{
.sample.not.NA(x)
}
\arguments{
\item{x}{numeric. A vector containing values to resampling.}
}
\value{
vector
}
\description{
Intern function to sample vectors with non-NA values
}
