% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspat_str.R
\name{.str.sample}
\alias{.str.sample}
\title{Vectorized structured sample}
\usage{
.str.sample(x, sp, resu, fr_prob)
}
\arguments{
\item{x}{vector containing sample size (i.e. richness) in the first element
and probabilities on the remaining}

\item{sp}{elements to be sampled}

\item{resu}{vector of results, must be the same length of sp}

\item{fr_prob}{Frequency or probability of each sp element. This will be multiplied by the
probability vector}
}
\value{
vector of sampled sp elements
}
\description{
Vectorized structured sample
}
\author{
Neander Marcel Heming
}
\keyword{internal}
