% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_check.R
\name{fit.memory}
\alias{fit.memory}
\title{Function to evaluate if the rasters generated in the function fit on RAM memory}
\usage{
fit.memory(x, n = 1)
}
\arguments{
\item{x}{SpatRaster}

\item{n}{positive integer. The number of copies of \code{x} that are needed}
}
\value{
logical
}
\description{
Function to evaluate if the rasters generated in the function fit on RAM memory
}
\author{
Neander Marcel Heming and Gabriela Alves-Ferreira
}
