% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootspat_str.R
\name{fr2prob}
\alias{fr2prob}
\title{Adjust probability of sampling based on frequency of occurrences.}
\usage{
fr2prob(x)
}
\arguments{
\item{x}{SpatRaster. A presence-absence SpatRaster.}
}
\value{
numeric vector
}
\description{
This function is to be used for adjusting probability of sampling a species to
get the frequency of sampled occurrences closer to the observed frequencies.
}
\examples{
\dontrun{
library(SESraster)
library(terra)
# creating random species distributions
f <- system.file("ex/elev.tif", package="terra")
r <- rast(f)
set.seed(510)
r10 <- rast(lapply(1:18,
                function(i, r, mn, mx){
                  app(r, function(x, t){
                    sapply(x, function(x, t){
                       x<max(t) & x>min(t)
                    }, t=t)
                  }, t=sample(seq(mn, mx), 2))
                }, r=r, mn=minmax(r)[1]+10, mx=minmax(r)[2]-10))

names(r10) <- paste("sp", 1:nlyr(r10))
fr2prob(r10)
# raw frequencies
subset(terra::freq(r10), value==1)[,"count"]
}
}
