% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ext_data.R
\name{load_ext_data}
\alias{load_ext_data}
\title{Load SESraster external datasets}
\source{
Rosauer, 2017. Available on: \href{https://github.com/DanRosauer/phylospatial/tree/master/PhyloEndemism_in_R/Tree\%20Frog\%20Data/}{Github}

IUCN. 2022. The IUCN Red List of Threatened Species (spatial data). Version 2022-1. \href{https://www.iucnredlist.org}{IUCN}
}
\usage{
load_ext_data(x = "spp_sites")
}
\arguments{
\item{x}{dataset to be loaded}
}
\value{
SpatRaster
}
\description{
This function loads external datasets available at extdata package folder
}
\details{
These are the available datasets:
\itemize{
\item spp_sites: a SpatRaster with randomly generated presence-absence data for five species.
}
}
