% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ext_data.R
\name{load_ext_data}
\alias{load_ext_data}
\title{Load SESraster external datasets}
\usage{
load_ext_data(x = "spp_sites")
}
\arguments{
\item{x}{dataset to be loaded}
}
\value{
SpatRaster object
}
\description{
This function loads external datasets available at extdata package folder
}
\details{
These are the available datasets:
\itemize{
\item spp_sites: a SpatRaster with randomly generated presence-absence data for five species.
}
}
\examples{
# load random species distributions
library(SESraster)
library(terra)

r <- load_ext_data()

plot(r)

}
